+++ test/device_tests.jl
@@ -524,13 +524,15 @@ const POINTWISE_BODIES = (
     :_pair_backward_remote_point!, :_pair_forward_local_point!,
     :_pair_forward_remote_point!, :_pair_select_point!,
     :_primitives_ideal_point!, :_primitives_stiffened_point!, :_rate_point!,
+    :_reciprocal_interior_point!, :_reciprocal_point!,
     :_rho_sensor_point!, :_ring_accum_point!, :_ring_pack_point!, :_rk_point!,
-    :_scale_grad_point!, :_shell_ring_point!, :_slip_wall_point!,
+    :_scale_grad_point!, :_scale_interior_point!, :_shell_ring_point!,
+    :_slip_wall_point!,
     :_species_bound_point!, :_species_diffusivity_point!, :_strain_mag_point!,
     :_subtract_div_point!,
     :_subtract_jac_div_point!, :_tag_delta4_point!, :_tag_gradient_point!,
     :_tag_rho_point!, :_tag_sensor_point!, :_tag_vorticity_point!,
-    :_zero_component_point!)
+    :_volume_weight_point!, :_weighted_blend_point!, :_zero_component_point!)
 
 @testset "pointwise bodies stay inside the splat budget" begin
     # `_point_kernel!` calls `body!(args..., i, j, k)`. Julia expands a